+++ packages/react-native-eyecandy-icons/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @nomada-sh/react-native-eyecandy-icons
 
+## 0.0.6
+
+### Patch Changes
+
+- Fixed Icon `color` prop literal string union autocomplete.
+- Updated dependencies
+  - @nomada-sh/react-native-eyecandy-theme@0.0.4
+
 ## 0.0.5
 
 ### Patch Changes