+++ packages/react-native-eyecandy-theme/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @nomada-sh/react-native-eyecandy-theme
 
+## 0.0.4
+
+### Patch Changes
+
+- Added `warning`, and `success` colors to text color choices.
+
 ## 0.0.3
 
 ### Patch Changes